+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Add `regionId` to the `simulationPayload`.
+
 ## [1.15.0] - 2020-08-13
 
 ### Added
-
 - Categories and categoriesIds to compatibility layer from all trees.
 
 ## [1.14.1] - 2020-08-06